only sent updates if there's a difference from the last tweet
Previously, we'd maintain a rolling list of updates, and send a tweet with
the union of all those updates. In some case, this resulted in a tweet where
the number of spots was the same in both tweets.

Instead, we build a new structure called `IdMapSet`. It's like a `MapSet`,
but rather than being unique by identity, it's unique by a function. This
helps us maintan the list of garages, and lets us easily calculate the
differences between the previous garage state and the current garage state.